+++ Tests/Test_CrossAssemblyInjections_SingletonCycle.swift
@@ -36,6 +36,7 @@ class Test_CrossAssemblyInjections_SingletonCycle: XCTestCase {
                     self.testChildObjectAssembly.testChildObjectPrototype,
                     self.testChildObjectAssembly.testChildObjectPrototype
                 ]
+                return $0
             }
         }
     }
@@ -49,12 +50,14 @@ class Test_CrossAssemblyInjections_SingletonCycle: XCTestCase {
         var testChildObject: TestChildObject {
             return define(init: TestChildObject()) {
                 $0.parentObject = self.testObjectAssembly.testObject
+                return $0
             }
         }
         
         var testChildObjectPrototype: TestChildObject {
             return define(scope: .prototype, init: TestChildObject()) {
                 $0.parentObject = self.testObjectAssembly.testObject
+                return $0
             }
         }
     }